+++ R/c2z.R
@@ -354,6 +354,7 @@ ZoteroFormat <- \(data = NULL,
                        "creators",
                        "relations")
   double.items <- c("version", "mtime")
+  dataframe.items <- c("relations", "tags")
   list.items <- c("collections", "relations", "tags")
 
   # Check if metadata and not in a data frame
@@ -376,6 +377,7 @@ ZoteroFormat <- \(data = NULL,
       # Add to list if element is a data frame
       ## Make certain list.items is in a list
       if (is.data.frame(x) | names %in% list.items) {
+        if (names %in% dataframe.items) x <- as.data.frame(x)
         if (!all(is.na(x))) x <- list(x)
         # Make certain double.items are double
       } else if (names %in% double.items) {